Show component code frame, if available
Summary:
## Overview
This change adds code frames for component stacks if it differs from the call stack frame.

## Background
After adding native stack based stack frames, which we already symbolicate, we had the capability to show a code frame for component stacks as well as the stack frame (if they differ). 

However, this usually wasn't that useful, because the native component stack was unlikely to be more useful than the component stack (see the comparisons below for key errors).

## Owner stacks 

With owner stacks the component frame is a lot more useful and in many cases are better at showing the location than the call stack frame. 

## Example Screens

Before:
{F1974436645} 

After (with owner stacks):
{F1974436723}  

Changelog:
[General][Added] - Add owner stack code frames to LogBox

Differential Revision: D68285627