Add prepare-ios script (#48791)
Summary:

This change moves the same scripts we have to prepare the HelloWorld project to RNTester.

This is something we forgot to do when we were decoupling the reacrt-native from the community CLI.

This is also the base to start deprecating cocoapods and add more configuration steps for the project.

## Changelog:
[Internal] - Copy cli.js script from HelloWorld to RNTester

Reviewed By: cortinico

Differential Revision: D68325581