introduce Fantom.dispatchNativeStateUpdate
Summary:
changelog: [internal]

Introduces a mechanism to fake C++ state update from JavaScript for testing in Fantom.

One problem with approach in this diff is use of folly::dynamic. To make it work, folly::dynamic is no longer Android only.
iOS and other platforms are not paying extra cost for folly::dynamic being included because they all already have it.

The alternative approach would be to introduce a component specific interface to NativeFantom module so state in ScrollView, TextInput and Image can be manipulated. On the flip side, this is a very flexible solution.

In the future, we might want to introduce Fantom wrappers around ScrollView and TextInput that would be flow typed in tests. They would also make sure that changing ScrollView.contentOffset always triggers both onScroll and state update.

Differential Revision: D68418535